fix double float conversion
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@17865 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: baec79d1dc6b2f2e15950579929c683f861f67ec

+++ LHCb/Tf/TsaKernel/TsaKernel/SeedHit.h
@@ -1,4 +1,4 @@
-// $Id: SeedHit.h,v 1.4 2007-10-09 18:03:39 smenzeme Exp $
+// $Id: SeedHit.h,v 1.5 2008-05-29 10:56:06 smenzeme Exp $
 #ifndef SEEDHIT_H
 #define SEEDHIT_H 1
 
@@ -188,7 +188,7 @@ namespace Tf
 
       inline bool onTrack() const { return tfHit()->onTrack(); }
 
-      inline float length() const { return tfHit()->hit()->length(); }
+      inline double length() const { return tfHit()->hit()->length(); }
 
     public: // setters
